Generate scale tensors together with data tensors
Before this commit, on parameter change, MW generates the scale tensors before
it generates the data tensors. It first sends the scales, then waits for the
"request-wavetables" message and *then* generates the data tensors.

With this commit, they are both not generated before RT sent the realtime
parameters.

Consequently:

* /request-wavetables only transfers Tensor scale pointers if it's not caused
  by parameter change
* MW does not only transfer a Tensor3 to RT, but the whole WaveTable
* ADnoteParams now has 2 WaveTables (current/next) instead of one WaveTable
  containing two Tensors

Some of these code changes will be undone with the following commit, where no
more Tensor1s will be transferred from RT to MW.