Communication: Change which table AD fills
In "set-waves", AD note decides which table to fill - the "current" or
the "next" table. The prior formulation makes it ambigous, saying that
the criterion is both whether Tensor2 vs Tensor1 have been send and
whether a parameter has been changed. The latter formulation has been
removed, this is also because future commits will have non-parameter
changes that do fill the "next" table.